% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-model.R
\name{as_gwl_ts}
\alias{as_gwl_ts}
\title{Coerce to \code{gwl_ts}}
\usage{
as_gwl_ts(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Passed to methods.}
}
\value{
A \code{gwl_ts}.
}
\description{
Coerce to \code{gwl_ts}
}
