% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-model.R
\name{check_gwl_ts}
\alias{check_gwl_ts}
\title{Check that an object satisfies the \code{gwl_ts} contract}
\usage{
check_gwl_ts(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Argument name to use in error messages.}

\item{call}{Calling environment for error messages.}
}
\value{
\code{TRUE}, invisibly, on success; otherwise a classed error.
}
\description{
Unlike \code{\link[=validate_gwl_ts]{validate_gwl_ts()}} this always inspects the object even if it is not
classed as \code{gwl_ts}, which makes it useful inside readers.
}
