% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-model.R
\name{check_gwl_wells}
\alias{check_gwl_wells}
\title{Check that an object satisfies the \code{gwl_wells} contract}
\usage{
check_gwl_wells(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Argument name to use in error messages.}

\item{call}{Calling environment for error messages.}
}
\value{
\code{TRUE} invisibly on success.
}
\description{
Check that an object satisfies the \code{gwl_wells} contract
}
