% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gems_ger_wells_sample}
\alias{gems_ger_wells_sample}
\title{GEMS-GER sample well metadata}
\format{
A \link{gwl_wells} \code{sf} layer (EPSG:25832) with \code{well_id} and curated
static attributes (\code{surface_elevation}, \code{well_depth}, \code{aquifer_medium},
\code{pressure_state}, ...).
}
\source{
As \link{gems_ger_sample}.
}
\usage{
gems_ger_wells_sample
}
\description{
Well locations and static attributes for the wells in \link{gems_ger_sample}.
}
\keyword{datasets}
