% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-model.R
\docType{data}
\name{gwl_ts_optional}
\alias{gwl_ts_optional}
\title{Optional-but-recognised columns of a \code{gwl_ts} table}
\format{
An object of class \code{character} of length 1.
}
\usage{
gwl_ts_optional
}
\description{
Optional-but-recognised columns of a \code{gwl_ts} table
}
\keyword{internal}
