% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-model.R
\docType{data}
\name{gwl_ts_required}
\alias{gwl_ts_required}
\title{Required columns of a \code{gwl_ts} table}
\format{
An object of class \code{character} of length 5.
}
\usage{
gwl_ts_required
}
\description{
Required columns of a \code{gwl_ts} table
}
\keyword{internal}
