% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-model.R
\docType{data}
\name{gwl_wells_crs}
\alias{gwl_wells_crs}
\title{Required columns / properties of a \code{gwl_wells} layer}
\format{
An object of class \code{integer} of length 1.
}
\usage{
gwl_wells_crs
}
\description{
Required columns / properties of a \code{gwl_wells} layer
}
\keyword{internal}
