% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-correctiv.R
\name{lap_correctiv_meta}
\alias{lap_correctiv_meta}
\title{CORRECTIV groundwater dataset metadata}
\usage{
lap_correctiv_meta()
}
\value{
A list with \code{source}, \code{licence}, \code{citation}, \code{url}, \code{crs}, \code{variable}.
}
\description{
CORRECTIV groundwater dataset metadata
}
