% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gems-ger.R
\name{lap_gems_ger_meta}
\alias{lap_gems_ger_meta}
\title{GEMS-GER dataset metadata}
\usage{
lap_gems_ger_meta(version = "latest")
}
\arguments{
\item{version}{Version key (\code{"1.0"}) or \code{"latest"}.}
}
\value{
A list with \code{version}, \code{record}, \code{doi}, \code{concept_doi}, \code{licence},
\code{citation}, \code{url}, \code{crs} and \code{variable}.
}
\description{
GEMS-GER dataset metadata
}
\examples{
lap_gems_ger_meta()
}
