% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave.R
\name{lap_preset_names}
\alias{lap_preset_names}
\title{Output presets for lapidary graphics}
\usage{
lap_preset_names()
}
\value{
A character vector of preset names.
}
\description{
Output presets for lapidary graphics
}
