% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-correctiv.R
\name{lap_read_correctiv_wells}
\alias{lap_read_correctiv_wells}
\title{Read CORRECTIV well metadata as a \code{gwl_wells} layer}
\usage{
lap_read_correctiv_wells()
}
\value{
A \code{gwl_wells} \code{sf} layer in EPSG:25832.
}
\description{
Read CORRECTIV well metadata as a \code{gwl_wells} layer
}
