% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{lap_variants}
\alias{lap_variants}
\title{Supported visual variants}
\usage{
lap_variants()
}
\value{
A character vector: \code{"light"}, \code{"dark"}.
}
\description{
Supported visual variants
}
\examples{
lap_variants()
}
