% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-model.R
\name{validate_gwl_ts}
\alias{validate_gwl_ts}
\title{Validate a \code{gwl_ts} table}
\usage{
validate_gwl_ts(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly, if valid; otherwise an error is thrown.
}
\description{
Validate a \code{gwl_ts} table
}
