% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind-catalog.R
\name{lap_ind_trend_extremes}
\alias{lap_ind_trend_extremes}
\title{Trend of the annual minima and maxima}
\usage{
lap_ind_trend_extremes(data, value = gwl, date = "date", min_years = 10L, ...)
}
\arguments{
\item{data}{A one-series data frame.}

\item{value}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> level column.}

\item{date}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> date column.}

\item{min_years}{Minimum number of years; below this the outputs are \code{NA}.}

\item{...}{Ignored; absorbs arguments forwarded to other indicators by
\code{\link[=lap_indicators]{lap_indicators()}}.}
}
\value{
A one-row tibble: \code{ind_trend_min_slope}, \code{ind_trend_max_slope}.
}
\description{
Theil-Sen slope (level units per year) of the series of annual minimum and of
annual maximum levels. Comparing \code{ind_trend_min_slope} with the mean trend
shows whether the \emph{drought floor} is dropping faster than the average.
}
\seealso{
Other indicators: 
\code{\link{lap_ind_amplitude}()},
\code{\link{lap_ind_climate_signal}()},
\code{\link{lap_ind_drought}()},
\code{\link{lap_ind_drought_recovery}()},
\code{\link{lap_ind_extreme_months}()},
\code{\link{lap_ind_flashiness}()},
\code{\link{lap_ind_memory}()},
\code{\link{lap_ind_phase_regularity}()},
\code{\link{lap_ind_recession}()},
\code{\link{lap_ind_recharge_discharge}()},
\code{\link{lap_ind_rise_fall}()},
\code{\link{lap_ind_seasonal_amplitude}()},
\code{\link{lap_ind_seasonality_strength}()},
\code{\link{lap_ind_step_change}()},
\code{\link{lap_ind_trend}()},
\code{\link{lap_ind_trend_acceleration}()}
}
\concept{indicators}
