% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapidary-package.R
\docType{package}
\name{lapidary-package}
\alias{lapidary}
\alias{lapidary-package}
\title{lapidary: Data Wrangling, Analysis and Styled Visualisation of German Groundwater-Level Time Series}
\description{
Tools to ingest, normalise and analyse groundwater-level time series for Germany (starting with the GEMS-GER benchmark dataset) and to produce a shared, poster- and web-ready ggplot2 style scheme. The package owns the data model, caching layer, analysis primitives (reference periods, water years, per-well summaries, normalisation, trends, spatial hexagon aggregation) and, from a later milestone, reusable plot builders. Downstream packages (a Shiny website, a poster builder) consume lapidary.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mxnl/lapidary}
  \item Report bugs at \url{https://github.com/mxnl/lapidary/issues}
}

}
\author{
\strong{Maintainer}: Max Noelscher \email{trashmail.mn@posteo.de}

}
\keyword{internal}
